back/requirements.txt
asgiref==3.8.1
asttokens==3.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
debugpy==1.8.12
decorator==5.2.1
dj-rest-auth==7.0.1
Django==4.2.20
django-allauth==65.8.1
django-environ==0.12.0
django-extensions==3.2.3
djangorestframework==3.16.0
drf-yasg==1.21.10
exceptiongroup==1.2.2
executing==2.2.0
idna==3.10
importlib_metadata==8.5.0
inflection==0.5.1
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
pure_eval==0.2.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
zipp==3.21.0
